:coffee: Add functionality for multiple sources
Change variables `source` and `destination` to array of elements for each line in the input. If the value was a single file, it's an array with one element.

New function groups compile CSS and write to file. The property `file` of `render_options` is set to the given SASS/SCSS file at the beginning of the function instead of at the beginning of the script.

This new function is run with all the sources and destinations.